refactor: :hammer: address review comments
